[PLAT-748] refactor averages component
